Publish both object_id and entity_id (#40)
Fixes #32 

This updates the index mapping to support both the `object_id` and the `entity_id`. Previously, we were publishing the `object_id` in the `entity_id` field, which was incorrect.

## Previous:
```json
{
  "_id": "...",
  "_source": {
     "domain": "switch",
     "entity_id": "living_room_light",
  }
}
```


## Updated:
```json
{
  "_id": "...",
  "_source": {
     "domain": "switch",
     "object_id": "living_room_light",
     "entity_id": "switch.living_room_light",
  }
}
```

As a result of this change, the previous index mappings are incompatible. To mitigate this, this PR also introduces a version string to the index template, index alias, and indices that are created. After upgrading, existing indices will no longer be written to, but the data will persist. New indices will be created to handle the updated mapping, and the numbering will restart at `hass-index-v1-000001`.

In my testing, existing Kibana index patterns are still valid, but require a refresh. Visualizations will likely need to be reworked if you are relying on the old `entity_id`.