Convert timestamp to UTC if time zone is missing
Some events in home assistant don't have a timezone information in time_fired and use local timezone, other events use UTC. When events arrive to elasticsearch, it treats all of them as UTC events. As a result, if ha runs in EST timezone device_tracker events are lagging 4-5 hours behind when displayed in kibana.

This commit checks if timezone is present and if it's not, converts the timestamp from local time to UTC.