Remove manage_ilm from docs (#529)
The docs did have manage_ilm in the docs when creating and API key but it seems the scripts itself do not use this anymore. This is relevant in the context of sending data to a Serverless instance where ILM does not exist.

Also see https://github.com/legrego/homeassistant-elasticsearch/issues/499